Implemented authorize in user routes

+++ src/api/routes/user.route.ts
@@ -1,6 +1,7 @@
 import { Router } from "express";
 import multer from "multer";
 import authenticate from "../../middlewares/authenticate.middleware";
+import authorize from "../../middlewares/authorize.middleware";
 import {
   list as listValidation,
   create as createValidation,
@@ -16,6 +17,7 @@ const router = Router({ mergeParams: true });
 router.get(
   "/",
   authenticate,
+  authorize("list", "user"),
   listValidation,
   UserController.list
 );
@@ -23,6 +25,7 @@ router.get(
 router.post(
   "/",
   authenticate,
+  authorize("create", "user"),
   upload.single("image"),
   createValidation,
   UserController.create
@@ -30,6 +33,7 @@ router.post(
 
 router.get(
   "/:id",
+  authorize("read", "user"),
   authenticate,
   UserController.read
 );
@@ -37,6 +41,7 @@ router.get(
 router.put(
   "/:id",
   authenticate,
+  authorize("update", "user"),
   upload.single("image"),
   updateValidation,
   UserController.update
@@ -45,6 +50,7 @@ router.put(
 router.put(
   "/change-password/:id",
   authenticate,
+  authorize("changePassword", "user"),
   changePasswordValidation,
   UserController.changePassword
 );
@@ -52,6 +58,7 @@ router.put(
 router.delete(
   "/:id",
   authenticate,
+  authorize("delete", "user"),
   UserController.remove
 );
 
@@ -59,6 +66,7 @@ router.post(
   "/delete-by-ids",
   authenticate,
   deleteByIdsValidation,
+  authorize("delete", "user"),
   UserController.deleteByIds
 );
